test(server): split edge-case suite into focused files; apply formatting; all tests green
